backend/requirements-medical.txt
pytesseract==0.3.10
opencv-python==4.8.1.78
Pillow==10.1.0
pdfplumber==0.9.0
PyMuPDF==1.23.8
requests==2.31.0
numpy==1.24.3